The standard user domain's 'save' method already has· an encode() method call in it - the password gets encode()d· twice, and one can never log in with one's password.
Maybe this should stay in these files, and the encode()ing in the·
User domain should be removed.  These files have 'salt'
mechanisms in them, which is missing from the user domain,
although that could be added there perhaps?

In any event, I have to modify these files to get basic
self registration and subseqent password authentication to work.